added images and fixed the style of essentials and added dynamic links for redirection
