--- src/components/home/FeaturedEssentialsRow.css
@@ -1,8 +1,0 @@
-/* FeaturedEssentialsRow.css */
-.hover-zoom img {
-  transition: transform 0.3s ease;
-}
-
-.hover-zoom:hover img {
-  transform: scale(1.1);
-}
